Classify Resource::Step for each status
