--- lib/turnip_formatter/resource/step.rb
@@ -1,23 +1,0 @@
-require 'forwardable'
-
-module TurnipFormatter
-  module Resource
-    class Step
-      extend Forwardable
-      def_delegators :@raw, :keyword, :text, :line, :argument
-
-      attr_reader :example
-      attr_accessor :status
-
-      #
-      # @param  [RSpec::Core::Example]  example
-      # @param  [Turnip::Node::Step]    raw
-      #
-      def initialize(example, raw)
-        @example = example
-        @raw = raw
-        @status = :passed
-      end
-    end
-  end
-end
