+++ lib/turnip_formatter/resource/step/pass.rb
@@ -1,0 +1,13 @@
+require 'turnip_formatter/resource/step/base'
+
+module TurnipFormatter
+  module Resource
+    module Step
+      class Pass < Base
+        def status
+          :passed
+        end
+      end
+    end
+  end
+end
