+++ lib/turnip_formatter/resource/step/unexecute.rb
@@ -1,0 +1,13 @@
+require 'forwardable'
+
+module TurnipFormatter
+  module Resource
+    module Step
+      class Unexecute < Base
+        def status
+          :unexecute
+        end
+      end
+    end
+  end
+end
